Finish implementation of UserLocation plugin
It now scans a set of user location databases (uses Config::Any,
so they can be in JSON, Yaml, Perl, XML and INI formats).

The location is only updated if it has changed
from the last time check_location was run.

Right now, we check the location each 60 seconds and immediatly after connect.

Signed-off-by: Pedro Melo <melo@simplicidade.org>